feat: make ruby useful (#43)
This introduces a useful set of code for Ruby. This is very much a spike, there are still some potential leaks in this code and there's no build system to actually emit the artefacts we actually need to consume this in another library. But the logic appears to be correct overall and this is a useful spike to building the other ffi consumers